ci: Adding Jenkinsfile
Adding a Jenkinsfile which automates the build, test, and deployment
of the application. It leverages two executors, one for Kaniko to build
the images without having to do Docker-in-Docker, and a python executor
which comes preinstalled with the requirements and other tools such as
pylint, pyright if they are needed. These are configured as kubernetes
cloud executors inside Jenkins.